fix: build docker issues fixed (compare page)
